+++ tests/testthat/to_style/packrat/no_style.R
@@ -1,0 +1,3 @@
+box::use(stringr)
+
+box::use(tidyr)