+++ NEWS.md
@@ -1,6 +1,7 @@
 # box.linters (development version)
 
-* Add a switch to `box_alphabetical_calls_linter()` to set verbosity of lint report.
+* Less verbose `box_alphabetical_calls_linter()`. Reports only the first out-of-place function.
+* Added styling functions for `box::use()` calls.
 * [Bug fix] Allow multiple `box::use(pkg)` calls (#111)
 
 # box.linters 0.9.1